Use API directly rather than calling out to `cf` cli
The previous proof of concept relied on 'shelling out' to cf commands as
a workaround for limitations with concurrent calls of the CliConnection
methods. This however had several drawbacks:
  * It relied on the cf command appearing in the user's path
  * It was very slow (30s+ to spin up the tunnel)
  * It was difficult to reason about errors reported back from cli

This is a re-write of the plugin that uses the CF API directly to
create and stage the required tunnel application, and configures the
tunneling using the Go ssh library.

The result is a significant speedup in deploying the tunnel and more
flexibility over retrying connections which leads to fewer connection
based errors (on our flaky network).