Add basic Travis config
Run `go vet` to detect simple errors and `go test` to see if it builds
successfully. `go test` doesn't need any test files in order to fail the
build when the code is syntactically incorrect.

We need to specify our own install steps because we need `dep` and Travis
doesn't run `go get` automatically when you have a `Makefile` present.

- https://docs.travis-ci.com/user/languages/go/#Dependency-Management

I've enabled Travis for this repo.