Get a new access token if expired
There was an issue with access tokens expiring and causing ephemeral
apps to be left behind[1,2].

Unfortunately, the Plugin API does not expose the refresh token[3], so
we work around this by using the `cf oauth-token` system command.

Tests have not been added for the JWT validation. This is because:

* if any aspect of it fails early the default is to refresh the token
* this is intended as a temporary workaround due to the issue of the
  Plugin API not exposing the refresh token[3]

[1] https://github.com/alphagov/paas-cf-conduit/issues/18
[2] https://github.com/alphagov/paas-cf-conduit/issues/26
[3] https://github.com/cloudfoundry/cli/issues/1372