Handle credentials more flexibly
Instead of a fix credentials structure we parse everything into a
map[string]interface{} and try to identify common fields with multiple names
(e.g. the host can be in host or hostname field).

This change will allow to pass the full credentials structure to the app, as
previously only the common fields were passed in VCAP_SERVICES.

Also thix fixes an issue where the uri and jdbc uri fields weren't updated with
localhost and the local port.