Add ssh keepalive to stop servers from hanging up
As per [alphagov/paas-cf-conduit#35](https://github.com/alphagov/paas-cf-conduit/issues/35)
`cf conduit` doesn't send any keepalive packets. This means if both the
client and the server are idle, the server may decide that the client
has gone away and close the connection.

Adding a background loop that sends an empty packet over the ssh
connection every (hardcoded for now) 2 seconds prevents this.

I've tested this manually by connecting to `pgcli` to a postgres
database and then ignoring it for 10 mins - this used to result in
confusing hanging behaviour and weird errors, with these changes it
doesn't.

To be consistent with the rest of this codebase I haven't written any
tests.