app: refactor, adding ability to re-use existing app
merge separate loops into initServiceBindings()' iteration of the
VcapServices so we can be sure we're using the same iteration logic
for checking service names, checking "program" correspondence and
checking binding information for all our requested services is
found. addition of service *name* checking in this phase means we
will no longer be exposing credentials of any non-requested
services that happen to also be bound to the app.

move initServiceBindings() call to SetupTunnels() as it is common
between deployed and existing apps.

create PrepareForExistingApp() which is responsible for fetching
information from the "existing app" equivalent to that gathered
during DeployApp()