Add option to (re-)use existing app, introduce tests
* app: refactor, adding ability to re-use existing app

merge separate loops into initServiceBindings()' iteration of the
VcapServices so we can be sure we're using the same iteration logic
for checking service names, checking "program" correspondence and
checking binding information for all our requested services is
found. addition of service *name* checking in this phase means we
will no longer be exposing credentials of any non-requested
services that happen to also be bound to the app.

move initServiceBindings() call to SetupTunnels() as it is common
between deployed and existing apps.

create PrepareForExistingApp() which is responsible for fetching
information from the "existing app" equivalent to that gathered
during DeployApp()

* cli: add --existing-app option

* cli: deprecate --reuse in favour of --no-delete

--reuse was already not a great description of what the flag
does, but it becomes particularly confusing with the addition
of --existing-app. luckily it was always a "hidden" option.

* cli: require --app-name when --existing-app is used, unhide option

need to move generation of random name so that we can detect
unset values

* client.Client: use gocfclient's client interface

* tool dependencies: add ifacemaker and counterfeiter

* client.Client: refactor to use as interface

rework construction and move RefreshAccessToken() procedure
in-object to reduce the number of fields we need to expose
through interface methods

* client.Credentials: abandon attempt at using this via an interface

use via an interface causes way too many json-unmarshaling
complexities, and the struct's methods are quite self-
contained, with only one mutating method. testing against
a fake implementation isn't really any easier than using
a real one.

* conduit.App: add tests

both "external" and "internal" tests, the latter just to allow
us to target initServiceBindings() more efficiently

* docs: providing documentation for the new flags

---------

Co-authored-by: Panos Xynos <panos.xynos@digital.cabinet-office.gov.uk>